Backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pilloW
django-jazzmin